refactor: move Toast paths and provider sequence
