+++ components/shared/Toast/index.tsx
@@ -2,6 +2,6 @@ import { Toast } from "./Toast";
 
 export * from "./Toast";
 export * from "./useToast";
-export * from "./CtxToastQueue";
+export * from "./ToastQueueContext";
 
 export default Toast;
